Fixes #28763 - split ChartBox and StatisticsChartsList
