+++ webpack/assets/javascripts/react_app/components/ChartBox/ChartBox.fixtures.js
