+++ webpack/assets/javascripts/react_app/components/ChartBox/ChartBox.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ChartBox from './ChartBox';
+import ChartBox from './';
 import mockStoryData from './ChartBox.fixtures';
 import Story from '../../../../../stories/components/Story';
