+++ webpack/assets/javascripts/react_app/components/ChartBox/index.js
