+++ webpack/assets/javascripts/react_app/components/StatisticsChartsList/StatisticsChartsList.fixtures.js
