+++ webpack/assets/javascripts/react_app/components/StatisticsChartsList/StatisticsChartsListStyles.scss
