+++ webpack/assets/javascripts/react_app/components/componentRegistry.js
@@ -4,7 +4,7 @@ import ReactApp from '../ReactApp';
 import DonutChart from './common/charts/DonutChart';
 import BarChart from './common/charts/BarChart';
 import LineChart from './common/charts/LineChart';
-import StatisticsChartsList from './statistics/StatisticsChartsList';
+import StatisticsChartsList from './StatisticsChartsList';
 import PowerStatus from './hosts/powerStatus/';
 import NotificationContainer from './notifications/';
 import ToastsList from './ToastsList/';
@@ -24,7 +24,7 @@ import SearchBar from './SearchBar';
 import Layout from './Layout';
 import EmptyState from './common/EmptyState';
 import ComponentWrapper from './common/ComponentWrapper/ComponentWrapper';
-import ChartBox from './statistics/ChartBox';
+import ChartBox from './ChartBox';
 import ConfigReports from './ConfigReports/ConfigReports';
 import DiffModal from './ConfigReports/DiffModal';
 import { WrapperFactory } from './wrapperFactory';
